mcp-weaviate-manager/requirements.txt
weaviate-client==4.4.4
